refactor: move subscription endpoints to StripeApi
